requirements.txt
boto3==1.28.52
botocore==1.31.52
cryptography==41.0.4
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.37
matplotlib-inline==0.1.6
numpy==1.26.0
pandas==1.3.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.2
scikit-learn==1.3.1
scipy==1.11.2
sqlparse==0.4.4
tabulate==0.9.0
hopsworks==3.2.0
hsfs==3.2.0
xgboost==1.6.1
mlflow==2.7.1